requirements.txt
flask
flask-login
flask-wtf
flask-sqlalchemy
werkzeug
python-dotenv
qrcode[pil]
reportlab
APScheduler>=3.10